tests: separate unit and it tests
