+++ tests/unit/test_data/ghsa/GHSA-cvx8-ppmc-78hm.json
