+++ tests/unit/test_data/redhat/CVE-2022-1798.json
